feat: use faster / better maintained log library
At a high level, this is swapping logrus for zap. It also attempts to
simplify how we use logging, but this is coming from a TS person; I'm
not sure what the usual standards are for Go. My thought process was
to have a singular log file and type def, where before there seem to be
several

I've also removed most of the configuration options; we can add these
back if we need them, but since this is a new project minimal config
seemed like a good starting point

Signed-off-by: Rob Cresswell <robcresswell@users.noreply.github.com>