feat: add a build that uses boring crypto to run on fips enabled hosts
A new build has been added that uses CGO and the boring crypto build
flag to substitute go's default crypto library with boring crypto to
ensure the binary can run on FIPs enabled hosts. The fips build includes
the go symbol table in the binary (note other builds include -s to strip
it), this allows you to run `go tool nm anchore-ecs-inventory | grep
'_Cfunc__goboringcrypto_'` to verify that the binary was built with
boringcrypto.

A new docker image will also be built with the tag `v1.0.0-fips-amd64`
that includes the -fips binary.

Signed-off-by: Bradley Jones <bradley.jones@anchore.com>